gke/cluster: gke-create-gpu-config: add
Create gpu_config as part of node startup

This config stores the number of GPU partitions to be created on the
node. This configuration is consumed by the device plugin after the
drivers are installed, to partition the GPUs as desired

This change already merged in 1.19.

BUG:182178632

Change-Id: I6c2b996761e2d4ea1f94810b6d2d982efdb95b30

generate gpu_config.json for time-sharing

Manually tested by applying the CL image to nodes

Change-Id: Ic46476e51ba041fb30ddaa37b5932d6fc9c023c3

update gpu cconfiguration in node script

Change-Id: I21d66998b3dbf8a13bca37b4dc16acbaf90d9791