in-place: Move setup-inplace-master-pod to gke-setup-inplace and copy the manifests before cluster starts
Change-Id: I55e7fbf54e6683410d5268d6f2a65553e7cec97f

change mv inplace cache to copy inplace cache to init suffix. This change
fix an issue that causes inplace installed components to be restarted by
kubelet.

${KUBE_HOME}/inplace/in-place-status.yaml is a layer of cache for
inplace agent. We added inplace init-step a while back, and during the
init step, agent cached the state onto disk. Previously, before the agent
was started by systemd, we move the init cache to
${KUBE_HOME}/inplace/in-place-status.init.yaml which captured the init
step info. However, since we removed the cache from the agent's
perspective, the agent redownload all the manifests and restarted all the
components. In this cl, we kept the cache in the system so inplace agent
won't download all the manifests after start up.

Change-Id: I9ffb29fd18d8b13219af421a2d521dacd99b113f