Add smt settings to boot process.
GKE Sandbox will change SMT settings via threads-per-core.
The current status quo is that SMT is disabled via that
baking process. SMT can be enabled by the method in the CL
below. An attempt was made to remove this call and sync
with cluster-server with the preloader, which led to
omg/48248.

This change will allow the environment variable, GVISOR_ENABLE_SMT
to be passed at runtime via the kube-env set of environment
variables. The cluster-server will be able to enable
SMT on versions that support threads-per-core in this way.

Change-Id: Ie295b0564f36ce6e76656c8aa36f7c44acfbb216
(cherry picked from commit a42ae2096bc7498390cf88dd9ce4cd4ed88cb1fd)