Enable container-level seccomp in GKE
Seccomp rules that are specified for containers are applied to the
container inside gVisor. In general, seccomp rules for containers
running inside gVisor is unnecessary, since the container is already
isolated from the host kernel (which is what the seccomp rules are
trying to protect). However, user may want to add an extra layer of
defense in depth to the container.

Note that seccomp is still blocked in the API and will require further
changes to work end-to-end.

b/216692868

Change-Id: I165be2972bbef762525d8cfb0f2dbe337b2d611a