Get boot disk name dynamically for BFQ setup
Previously, the path to IO Scheduler hardcoded `sda` as the boot disk.
This does not work on NVME devices, since the naming of the boot disk
device is different. Instead of hardcoding the path, look up the path
dynamically based on where `/mnt/stateful_partition` is mounted (which is
where COS stateful partition is located.

BUG=b/240487453

Signed-off-by: David Porter <porterdavid@google.com>
Change-Id: Idbd0fff40257f3321db4e861959f144fbdd847d6

Fix BFQ Config for nodes with local SSD

The previous logic for detecting the boot device name did not work
correctly for machines with local SSD, resulting in BFQ not being
configured properly. This is because the previous logic assumed that the
only boot disk device had mounts of `/mnt/stateful_partition`. For local
SSD, this is not the case, since there can be multiple devices mounting
that directory because they are raided together via `mdadm` and a new
virtual device is created, `/dev/md0` representing the RAID array. See
https://b.corp.google.com/issues/299228620#comment1 as example.

This change adjusts the logic of detecting the boot disk device to use
`rootdev` as recommended by COS team, see
http://g/gci-users/SfDNMoSJpGc/Tc9oO6nYAwAJ?utm_medium=email&utm_source=footer.
Using rootdev, detects the correct boot disk device, even when using
local SSDs.

TESTED=http://b/299228620#comment18

BUG=299228620

Change-Id: I7f0522d89a706503983d8e882819f48c9d72a69c