Get boot disk name dynamically for BFQ setup
Previously, the path to IO Scheduler hardcoded `sda` as the boot disk.
This does not work on NVME devices, since the naming of the boot disk
device is different. Instead of hardcoding the path, look up the path
dynamically based on where `/mnt/stateful_partition` is mounted (which is
where COS stateful partition is located.

BUG=b/240487453

Signed-off-by: David Porter <porterdavid@google.com>
Change-Id: Idbd0fff40257f3321db4e861959f144fbdd847d6