Add core-tagging flag to GKE Sandbox instances.
Core Tagging/Core Scheduling is a security feature available in 5.14+
kernels. Core Tagging will allow SMT to be enabled in all GKE Sandbox
nodes while mitigating the risks of side-channel attacks. Pass the flag
to the config.toml so that we can enable core-tagging in the proper versions
in the control plane.

See: go/gvisor-core-scheduling

Effort=core-tags

Change-Id: Ib4448ae32f9ab006569b66130b67061277bf5f81