gvisor: Disable directfs flag by default.
As of right now, directfs is not enabled in runsc by default.
But that will change soon as we ship this feature to OSS users.
Keep GKE Sandbox on the old default until we have more confidence
about this feature/optimization.

--directfs flag was added to runsc in 20230306.0_RC01 release.
1.27.0-gke.300 was the first GKE release to contain that runsc release.
So it should be safe to reference '--directfs' runsc flag on 1.27+
release branch.

Change-Id: I06f52df58f59209e65e189e39f35fedfc76b3830