Add a flag to disable force detach behavior in kube-controller-manager
This is an abbreviated version of https://github.com/kubernetes/kubernetes/pull/120344

It changes the boolean plumbing to use a global package variable to
avoid conflicts.

The unit test is added in an isolated file, named starting with z_ so it
runs after other OSS unit tests which are not resilient to metrics
starting at any value other than 0.

Bug: b/272460654
Change-Id: I5437f3d2dc73f4f1a3782ba3afe5fdc8f93287e4