Add logic to admission controller for GPU/TPU annotations.
Add logic to set GPU/TPU annotations for gVisor for the admission
controller. We should set the annotations only if a user requests
a GPU/TPU resource in a container's resource section.

In addition, add logic in Validate calls to ensure we don't set
nvproxy (GPUs) and tpuproxy (TPUs) annotations unless there is a
resource requested and not just a user setting the annotation
themselves. This is because some security settings are modified
in gVisor when we use GPU/TPU and we should not allow users to
set those flags w/o actually using a GPU/TPU.

Change-Id: Ibe30cb02be3260f5df5f2c5aa4f5e3d785d6c76e
(cherry picked from commit 9c14a21b985268e09c0d23da0ad314ebe41d96ed)