Optimize systemctl calls in node initialization
For every `enable` action (without extra options),
`systemctl` internally runs a `daemon-reload` action afterwards.
This fact is mentioned in the `systemctl` man page
and is confirmed by the systemd source code - look for the
`daemon_reload` calls in the `verb_enable` function.

This change saves us n daemon-reload actions,
where n is the number of enabled services (n=8 at the time of writing).

Change-Id: I7f2b64b722d1e090e40ba1a4772d319a8f5ce4b6
Bug: 326573533