Add standard script to install installables.
Add python script to install installables. The script:
1) Downloads either a container installable w/ ctr or an AppPkg (read
  tarball). In the case of AppPkgs it verifies the hash.
2) Installs the installable.
   In the case of Containers, it runs the  container with the provided
   arguments.
   In the case of AppPkgs, the user may provide a path to dump the tarball
   or a set of paths to install individual files. AppPkgs are downloaded
   to a temporary directory and cleaned up after installation.
3) The script provides logic for preloading and not-preloading
   installables. When preloading, downloads happen as normal. When not
   preloading, no downloads occur, and in the case of AppPkgs, no
   installation occurs (it should have been done at preloding).

The following assumptions are made on installables:

- URLs are "rendered" outside of the script. (gcr.io/gke-release/... instead
  of {STORAGE_ENDPOINT}/...)
- Containers URLS are {remoteURL}@{digestAlgo}:digest
- AppPkg URLs are simply their remoteURL and they are verified by the
  digest. Digests Algos can be anything that python's hashlib supports.
- AppPkg archives should be unwrapped w/ 'tar -xzf [FILE]'

Outside of built-in pytools, the script requires ctr and
tar.

Change-Id: Ief47cc4fe23638e52894678a0d6e681cb87fea8f

Add installable tests to unit test job.

Add installable tests to unit test job. Technically these are
integration tests, but they will run quick enough on the unit test
job to justify just running them there instead of on a new job.

We'll limit container kind tests (for now) because "ctr" is not present
on the test-infa container. I'll attempt to add it and see if it is
possible to add container tests to these.

Change-Id: I82c42d28001341a447fbc4dd4d0c90857299262a