Install and run k8s_pki to dynamically configure certificates on KCP VMs.
Add installing in preload function as we'd like to always preload the binary to meet Apple's requirements. If for some reason not preloaded, it will also be downloaded as preload is called in the main function here.

Change-Id: I72cee050a61d622339c414766b6415225609c66c

Put "once" just after k8s_pki instead of "--once" as "once" is a subcommand.

b/331379766

Change-Id: I17b5a0df5e808f93bb7ecb49352a63b3921f503e