Add installable tests to unit test job.
Add installable tests to unit test job. Tests are written in python but
called in golang so that they run in the existing unit test job.

The tests accept a "--fake" argument which fakes things like GCE
metadata, ctr/containerd, and downloading from GCS bucket links.

The python tests can be run without --fake on a GCE VM to validate the
scripts end to end.

Change-Id: I8f603912c8c835f7cd4baf2f4f56d35a6b52a813