Placeholders: Unload gvnic during suspend
As per discussion with GCE networking team, it is recommended to unload
the gvnic module prior to suspend rather than unloading and reloading it
on resume (as was implemented previously). See
http://b/375051558#comment10 for additional context.

To implement the change, introduce a separate systemd unit which runs on
suspend (similar as the existing one done for resume). Prior to suspend,
unload the module, and then on resume, we will only load the module.

BUG=375051558

Change-Id: Ibb1e67ff7f628d747bfcf14c26ade0797d68b02a