Adding script to run installable presubmits.
Add a script and CloudBuild workflow to run installable
tests as presubmits. The script will configure a cloud
build workflow, which will use the cos-customizer tool.
The workflow will spin up a COS VM, run a script on it,
and create a VM image. The scripts will then cleanup
older VM images so that CI projects do not gain quota
from CI scripts.

This workflow is very similar to the one the preloader uses.

Eventually, users who are integrating their components into
installable workloads will be expected to add their tests to
this workflow as part of integration.

Change-Id: I60f0df78e37e24afa87a7c64869188eeb3b88661