Add installables to configure.sh
Add installable calls to configure.sh

In addition, add an e2e test for installables.

With these changes, we change the initial idea of how installables
are passed. Instead of multiple environment variables, we have
a single environment variable. The variable stores a json map
where the keys are component names (forced to be unique in EVE)
and values are a list of installable values. See the e2e test
to read the format.

This allows users to, in the common case, add an installable in
Google3 only.

Users may need to check if their installable exists. In this case,
they can check the installable environment variable for the existance
of their component.

Change-Id: I951d143d291a770bcae7a7325fb1efe9d4b04aa0