remove env var check for setting s2a iptable rule
This avoids master/node being in an incompatible state during feature
rollout.

Details: This check causes issue for the GKE-S2A feature rollout in
gke-metadata-server. E.g. the master will pick up the 'enableS2A' value
first after upgrade, and re-start the gke-metatata-server pod on all
nodes, with the feature on. But the node upgrade doesn't happen at the
same time, or may not happen for a long time. A workload pod would learn
from the gke-metadata-server that the feature is on, but when it talks
to GKE-S2A, it will fail because the iptable rule on the node hasn't
been updated properly.

Note: it is okay for this iptable rule to exist while nothing is
listening on it. Clients will check with gke-metadata-server first,
before talking to GKE-S2A.

Change-Id: Iedf84b2e95e8af97a09a7f702492b6fbd1e53508