Add stateful partition resize to placeholder resume script
This change attempts to run `resize-stateful-partition.service` during
the placeholder resume phase to resize filesystem in order to support
disk upsize for boltVM

Tested in GKE Sandbox with GCE Staging:
https://fusion2.corp.google.com/fd2d795a-3913-32a9-9170-1cf7da07c1f9

testing process: With the GKE version that we got from this change, we have created a custom ssg cl/802744639 following this doc go/custom-ssgs with smaller disk size (100 GB) and then we have modified the consumption_test.go to create a VM with disk size 1000 GB (cl/808244865).

From the test log we verified that the VM was created using boltVM and custom ssg that we provided https://screenshot.googleplex.com/8xRqez3wNCHXycA

We also verified that the disk was upsized to ~1000 GB from 100 GB using the cloud audit log.

Design doc: go/gke-bolt-upsize-disk-on-consume

Change-Id: I6540c524b418f92e32bb40b64a9b801bf2bda52f